refactor: reorganize code structure for better maintainability
- Move extractTweetId to src/utils/extractTweetId.ts
- Move formatTweet to src/utils/formatTweet.ts with helper functions
- Update src/actions/readTweet.ts to re-export utilities
- Split tests by responsibility:
  - src/test/extractTweetId.test.ts (23 tests)
  - src/test/formatTweet.test.ts (15 tests)
- Follow plugin-depin structure pattern
- All 38 tests still passing
